Backend/pyproject.toml
[project]
name = "cijene"
version = "0.1.0"
description = "Service for product pricing data by Croatian grocery chains."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "lxml>=5.4.0",
    "openpyxl>=3.1.5",
    "pydantic>=2.11.4",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.34.2",
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=4.2.0",
    "pyright>=1.1.400",
    "ruff>=0.11.10",
    "ty>=0.0.1a3",
]
